backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "fastapi-mail>=1.5.8",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqladmin>=0.21.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
